refactor: DOM String 상수화, Throttle View로 이동, bind Events 통합
