+++ src/js/__test__/search.test.js
@@ -1,6 +1,6 @@
 import validator from '../utils/validator.js';
 import APIManager from '../managers/APIManager.js';
-import { fetchData, errorData, parseData } from '../utils/constants.js';
+import { fetchData, errorData, parseData } from '../utils/mockData.js';
 
 describe('보고 싶은 영상을 검색 했을 때', () => {
   test('입력 없이 버튼을 눌렀다면 error를 throw한다.', () => {
