+++ src/js/utils/common.js
@@ -1,0 +1,11 @@
+export function throttleScroll() {
+  let throttle = null;
+  return function (callback) {
+    if (!throttle) {
+      throttle = setTimeout(() => {
+        throttle = null;
+        callback();
+      }, 1000);
+    }
+  };
+}
