+++ src/js/utils/mockData.js
@@ -1,0 +1,543 @@
+export const mockData = [
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 0,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 1,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 2,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 3,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 4,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 5,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 6,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 7,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 8,
+  },
+  {
+    url: 'https://i.ytimg.com/vi/ECfuKi5-Cfs/hq720.jpg?sqp=-oaymwEcCOgCEMoBSFXyq4qpAw4IARUAAIhCGAFwAcABBg==&rs=AOn4CLDvmIcX-TgdcH2g_Bd4AUxw6hjmvQ',
+    title: '[Playlist] 너무 좋은데 괜찮으시겠어요?',
+    channelName: 'essential;',
+    publishedDate: '2022년 3월 2일',
+    saveButton: '⬇ 저장',
+    id: 9,
+  },
+];
+
+export const fetchData = {
+  kind: 'youtube#searchListResponse',
+  etag: 'lQ-VEF9lgCYrZErL9nEFN1I1-_U',
+  nextPageToken: 'CAoQAA',
+  regionCode: 'KR',
+  pageInfo: {
+    totalResults: 604,
+    resultsPerPage: 10,
+  },
+  items: [
+    {
+      kind: 'youtube#searchResult',
+      etag: 'utwze-P6EJ5x-0heE9punV2ZcGQ',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'mQhgF7RoUCA',
+      },
+      snippet: {
+        publishedAt: '2021-10-14T19:06:42Z',
+        channelId: 'UCGmYJSYFM19VgwEQJsY12Dg',
+        title: '우아한테크코스 4기 온라인 설명회',
+        description:
+          '2022년 2월에서 11월까지 10개월 과정으로 진행하는 우아한테크코스 4기에 대한 설명회이다. 설명회는 우아한테크코스 교육 ...',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/mQhgF7RoUCA/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/mQhgF7RoUCA/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/mQhgF7RoUCA/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '박재성',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-10-14T19:06:42Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: '5ArA6FxNCDie_zvTs0C_J-heank',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'ytt37XkcHTU',
+      },
+      snippet: {
+        publishedAt: '2021-06-03T01:41:57Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 🎬vlog] Ep.4 그루밍의 하루 ☀️',
+        description:
+          '우테코로그 Ep.4 그루밍의 하루 ☀️ 돌아온 우테코 vlog의 첫 번째 주자는 3기 크루 중 가장 밝은 텐션을 자랑하는 그루밍입니다 : ) 잘 ...',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/ytt37XkcHTU/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/ytt37XkcHTU/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/ytt37XkcHTU/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-06-03T01:41:57Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: 'MgTCXMypeUICimjHgQasrMiz1Ig',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'nhQRaRgV19o',
+      },
+      snippet: {
+        publishedAt: '2021-09-09T04:04:59Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 인터뷰 챌린지] 🎙 코치에게 묻다! 코테뷰 #4 포비',
+        description:
+          "우아한테크코스의 크루들이 진행하는 인터뷰 챌린지입니다.   이번 우테코 인터뷰 챌린지는 '코치 인터뷰'로 진행되었습니다.",
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/nhQRaRgV19o/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/nhQRaRgV19o/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/nhQRaRgV19o/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-09-09T04:04:59Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: 'wT6nFuYCuRdM_6L427Rmrx-dZE4',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'HxzKg7V6r00',
+      },
+      snippet: {
+        publishedAt: '2019-10-05T01:19:06Z',
+        channelId: 'UCipvQqo32UmHkAAwhn-MfbQ',
+        title: '우아한테크코스 1기 합격 노하우 전격 공개. 비전공자 크루가 직접 들려주는 이야기 2편',
+        description:
+          '우아한테크코스 후기를 기록한 영상입니다. 이 영상에서는 우아한테크코스의 선발과정에 대하여 다룹니다. 1편 우아한테크코스 2기?',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/HxzKg7V6r00/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/HxzKg7V6r00/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/HxzKg7V6r00/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '개발왕루피',
+        liveBroadcastContent: 'none',
+        publishTime: '2019-10-05T01:19:06Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: 'hHO4w55-_FhCWs7lF6gARTbCDPI',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'Xm0SmqBGaBA',
+      },
+      snippet: {
+        publishedAt: '2020-01-08T07:36:40Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 🎬vlog] 우테코 1기가 2기에게 전하는 ✉️메세지',
+        description:
+          '2019년 12월 27일 우아한테크코스의 1기 크루들이 긴 항해  를 마치고, \u200d  수료식 \u200d  을 가졌습니다. 이 때 특별한 시간을 가져 ...',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/Xm0SmqBGaBA/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/Xm0SmqBGaBA/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/Xm0SmqBGaBA/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2020-01-08T07:36:40Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: 'lod3jLdLU-rYGR7AC8AYYZluSK4',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'ZQklkmFlYQI',
+      },
+      snippet: {
+        publishedAt: '2021-07-28T05:20:22Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 🎬vlog] Ep.5 우테코 프로젝트 비하인드 🎞',
+        description:
+          '우테코로그 Ep.5 우테코 프로젝트 비하인드 이번 우테코로그는 현재 크루들이 진행하고 있는 프로젝트 선정 비하인드 영상입니다 ...',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/ZQklkmFlYQI/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/ZQklkmFlYQI/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/ZQklkmFlYQI/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-07-28T05:20:22Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: 'Uh6GOEEuQV7wP5slyO4VQInl9eY',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'r1AWm783YeY',
+      },
+      snippet: {
+        publishedAt: '2021-09-09T03:30:44Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 인터뷰 챌린지] 🎙 코치에게 묻다! 코테뷰 #3 포코',
+        description:
+          "우아한테크코스의 크루들이 진행하는 인터뷰 챌린지입니다.   이번 우테코 인터뷰 챌린지는 '코치 인터뷰'로 진행되었습니다.",
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/r1AWm783YeY/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/r1AWm783YeY/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/r1AWm783YeY/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-09-09T03:30:44Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: '51SjOy2nssfj71Sxh3ix-1Ex8Eg',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'enINkqC7FAc',
+      },
+      snippet: {
+        publishedAt: '2021-07-14T05:16:07Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 인터뷰 챌린지] 🧡 프론트엔드 백엔드 서로에게 묻다! #1',
+        description:
+          '우아한테크코스의 크루들이 진행하는 인터뷰 챌린지입니다.   우테코 인터뷰 챌린지 이번 편은 백엔드와 프론트엔드 크루들이 평소에 ...',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/enINkqC7FAc/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/enINkqC7FAc/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/enINkqC7FAc/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-07-14T05:16:07Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: 'bfS-x6xKV14TESX5ieO6xEfY4vk',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'UGsM1vSTAno',
+      },
+      snippet: {
+        publishedAt: '2021-07-14T06:28:33Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 인터뷰 챌린지] 🧡 프론트엔드 백엔드 서로에게 묻다! #3',
+        description:
+          '우아한테크코스의 크루들이 진행하는 인터뷰 챌린지입니다.   우테코 인터뷰 챌린지 이번 편은 백엔드와 프론트엔드 크루들이 평소에 ...',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/UGsM1vSTAno/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/UGsM1vSTAno/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/UGsM1vSTAno/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-07-14T06:28:33Z',
+      },
+    },
+    {
+      kind: 'youtube#searchResult',
+      etag: 'K67nj_5cOKajD_nSPwOpiDXjWYY',
+      id: {
+        kind: 'youtube#video',
+        videoId: 'eRffd42sd4I',
+      },
+      snippet: {
+        publishedAt: '2021-11-29T08:06:30Z',
+        channelId: 'UC-mOekGSesms0agFntnQang',
+        title: '[우테코 🎬vlog] Ep. 8 👾우아한테크코스 3기 데모데이🧑\u200d💻👩\u200d💻',
+        description:
+          '우테코로그 Ep. 8 우아한테크코스 데모데이   지난 10월 29일, 크루들이 밤낮으로 고생하며 완성한 프로젝트의 최종 데모데이를 진행 ...',
+        thumbnails: {
+          default: {
+            url: 'https://i.ytimg.com/vi/eRffd42sd4I/default.jpg',
+            width: 120,
+            height: 90,
+          },
+          medium: {
+            url: 'https://i.ytimg.com/vi/eRffd42sd4I/mqdefault.jpg',
+            width: 320,
+            height: 180,
+          },
+          high: {
+            url: 'https://i.ytimg.com/vi/eRffd42sd4I/hqdefault.jpg',
+            width: 480,
+            height: 360,
+          },
+        },
+        channelTitle: '우아한Tech',
+        liveBroadcastContent: 'none',
+        publishTime: '2021-11-29T08:06:30Z',
+      },
+    },
+  ],
+};
+
+export const errorData = {
+  error: {
+    code: 400,
+    message: 'API key not valid. Please pass a valid API key.',
+    errors: [
+      {
+        message: 'API key not valid. Please pass a valid API key.',
+        domain: 'global',
+        reason: 'badRequest',
+      },
+    ],
+    status: 'INVALID_ARGUMENT',
+    details: [
+      {
+        '@type': 'type.googleapis.com/google.rpc.ErrorInfo',
+        reason: 'API_KEY_INVALID',
+        domain: 'googleapis.com',
+        metadata: {
+          service: 'youtube.googleapis.com',
+        },
+      },
+    ],
+  },
+};
+
+export const parseData = [
+  {
+    videoId: 'mQhgF7RoUCA',
+    publishedAt: '2021-10-14T19:06:42Z',
+    title: '우아한테크코스 4기 온라인 설명회',
+    url: 'https://i.ytimg.com/vi/mQhgF7RoUCA/mqdefault.jpg',
+    channelTitle: '박재성',
+  },
+  {
+    videoId: 'ytt37XkcHTU',
+    publishedAt: '2021-06-03T01:41:57Z',
+    title: '[우테코 🎬vlog] Ep.4 그루밍의 하루 ☀️',
+    url: 'https://i.ytimg.com/vi/ytt37XkcHTU/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+  {
+    videoId: 'nhQRaRgV19o',
+    publishedAt: '2021-09-09T04:04:59Z',
+    title: '[우테코 인터뷰 챌린지] 🎙 코치에게 묻다! 코테뷰 #4 포비',
+    url: 'https://i.ytimg.com/vi/nhQRaRgV19o/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+  {
+    videoId: 'HxzKg7V6r00',
+    publishedAt: '2019-10-05T01:19:06Z',
+    title: '우아한테크코스 1기 합격 노하우 전격 공개. 비전공자 크루가 직접 들려주는 이야기 2편',
+    url: 'https://i.ytimg.com/vi/HxzKg7V6r00/mqdefault.jpg',
+    channelTitle: '개발왕루피',
+  },
+  {
+    videoId: 'Xm0SmqBGaBA',
+    publishedAt: '2020-01-08T07:36:40Z',
+    title: '[우테코 🎬vlog] 우테코 1기가 2기에게 전하는 ✉️메세지',
+    url: 'https://i.ytimg.com/vi/Xm0SmqBGaBA/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+  {
+    videoId: 'ZQklkmFlYQI',
+    publishedAt: '2021-07-28T05:20:22Z',
+    title: '[우테코 🎬vlog] Ep.5 우테코 프로젝트 비하인드 🎞',
+    url: 'https://i.ytimg.com/vi/ZQklkmFlYQI/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+  {
+    videoId: 'r1AWm783YeY',
+    publishedAt: '2021-09-09T03:30:44Z',
+    title: '[우테코 인터뷰 챌린지] 🎙 코치에게 묻다! 코테뷰 #3 포코',
+    url: 'https://i.ytimg.com/vi/r1AWm783YeY/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+  {
+    videoId: 'enINkqC7FAc',
+    publishedAt: '2021-07-14T05:16:07Z',
+    title: '[우테코 인터뷰 챌린지] 🧡 프론트엔드 백엔드 서로에게 묻다! #1',
+    url: 'https://i.ytimg.com/vi/enINkqC7FAc/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+  {
+    videoId: 'UGsM1vSTAno',
+    publishedAt: '2021-07-14T06:28:33Z',
+    title: '[우테코 인터뷰 챌린지] 🧡 프론트엔드 백엔드 서로에게 묻다! #3',
+    url: 'https://i.ytimg.com/vi/UGsM1vSTAno/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+  {
+    videoId: 'eRffd42sd4I',
+    publishedAt: '2021-11-29T08:06:30Z',
+    title: '[우테코 🎬vlog] Ep. 8 👾우아한테크코스 3기 데모데이🧑\u200d💻👩\u200d💻',
+    url: 'https://i.ytimg.com/vi/eRffd42sd4I/mqdefault.jpg',
+    channelTitle: '우아한Tech',
+  },
+];
